requirements.txt
PyAutoGUI==0.9.52
idle_time==0.1.0
progressbar33==2.4
